+++ packages/date/index.js
@@ -3,9 +3,9 @@ const getDateHourMinutes = require('../../utils/date/getDateHourMinutes')
 const getDayMonthYear = require('../../utils/date/getDayMonthYear')
 const getFullRelativeTime = require('../../utils/date/getFullRelativeTime')
 const getHourMinutes = require('../../utils/date/getHourMinutes')
-const getIsoRelativeTime = require('../../utils/date/getIsoRelativeTime')
+const getISOAndRelativeTime = require('../../utils/date/getISOAndRelativeTime')
 const getRelativeTime = require('../../utils/date/getRelativeTime')
-const getYearMonthDay = require('../../utils/date/getYearMonthDay')
+const getYearMonthDay = require('../../utils/date/toYearMonthDay')
 const revYearMonthDay = require('../../utils/date/revYearMonthDay')
 
 // export packages
@@ -14,7 +14,7 @@ module.exports = {
 	getDayMonthYear,
 	getFullRelativeTime,
 	getHourMinutes,
-	getIsoRelativeTime,
+	getISOAndRelativeTime,
 	getRelativeTime,
 	getYearMonthDay,
 	revYearMonthDay,