+++ utils/date/getCurrentISO.js
@@ -1,0 +1,5 @@
+const getISO = require('./toISO')
+const getNow = require('./getNow')
+
+// get MS value of current time
+module.exports = () => getISO(getNow())