+++ utils/date/getMs.js
@@ -1,0 +1,2 @@
+// get MS value of timestamp
+module.exports = (date) => date.valueOf()