+++ utils/date/getNow.js
@@ -1,0 +1,4 @@
+const { DateTime } = require('luxon')
+
+// get iso date with relative years
+module.exports = () => DateTime.now()