+++ utils/date/parseISO.js
@@ -1,0 +1,4 @@
+const { DateTime } = require('luxon')
+
+// parse ISO format
+module.exports = (iso) => DateTime.fromISO(iso)