+++ utils/date/toISO.js
@@ -1,0 +1,2 @@
+// get ISO format
+module.exports = (date) => date.toUTC().toISO()